+++ atrvjr_drivers/rflex/src/atrvjr_node.cc
@@ -52,10 +52,10 @@ class ATRVJRNode {
         tf::TransformBroadcaster broadcaster; ///< Transform Broadcaster (for odom)
 
         bool isSonarOn;
-        float last_distance, last_bearing;
-        float x_odo, y_odo, a_odo;
+        double last_distance, last_bearing;
+        double x_odo, y_odo, a_odo;
         bool initialized;
-        float first_bearing;
+        double first_bearing;
         int prev_bumps;
 
         void publishOdometry();
@@ -218,8 +218,8 @@ void ATRVJRNode::publishOdometry() {
         return;
     }
 
-    float distance = driver.getDistance();
-    float true_bearing = angles::normalize_angle(driver.getBearing());
+    double distance = driver.getDistance();
+    double true_bearing = angles::normalize_angle(driver.getBearing());
 
     if (!initialized) {
         initialized = true;
@@ -228,9 +228,9 @@ void ATRVJRNode::publishOdometry() {
         y_odo = 0;
         a_odo = 0*true_bearing;
     } else {
-        float bearing = true_bearing - first_bearing;
-        float d_dist = distance-last_distance;
-        float d_bearing = bearing - last_bearing;
+        double bearing = true_bearing - first_bearing;
+        double d_dist = distance-last_distance;
+        double d_bearing = bearing - last_bearing;
 
         if (d_dist > 50 || d_dist < -50)
             return;
@@ -273,7 +273,7 @@ void ATRVJRNode::publishOdometry() {
 
     //set the velocity
     odom.child_frame_id = "base_link";
-    float tvel = driver.getTransVelocity();
+    double tvel = driver.getTransVelocity();
     odom.twist.twist.linear.x = tvel*cos(a_odo);
     odom.twist.twist.linear.y = tvel*sin(a_odo);
     odom.twist.twist.angular.z = driver.getRotVelocity();
